add global var (50%)

+++ test_prog/01-simple_test/libtest.c
@@ -1,8 +1,8 @@
 #include "libtest.h"
 
-int VERSION = 3;
+long VERSION = 3;
 
-int libtest_version() {
+long libtest_version() {
 	return VERSION;
 }
