+++ test_prog/01-simple_test/libtest.h
@@ -1,7 +1,7 @@
 #ifndef LIB_TEST_H_
 #define LIB_TEST_H_
 
-int libtest_version();
+long libtest_version();
 
 int libtest_add(int a, int b);
